feat: pass vaults to bottom sheet
